dnl1/requirements.txt
# Core dependencies
aiohttp>=3.8.0
aiofiles>=0.8.0
asyncio>=3.4.3
uvloop>=0.16.0; platform_system != "Windows"  # uvloop for Unix systems only
uvloop
rich>=12.0.0
requests>=2.28.0
urllib3>=2.0.0
pathlib>=1.0.1

# Protocol handlers
yt-dlp>=2023.10.0
libtorrent
aioftp>=0.21.0
asyncssh>=2.12.0
aiowebdav>=0.5.0
paramiko>=2.12.0
m3u8>=3.3.0

# Web interface
fastapi>=0.95.0
uvicorn>=0.21.0
websockets>=10.4
jinja2>=3.1.0
aiohttp-session>=2.12.0
python-multipart>=0.0.5
starlette>=0.26.0

# Performance and monitoring
psutil>=5.9.0
memory-profiler>=0.61.0
py-cpuinfo>=9.0.0

# Data handling and storage
pyyaml>=6.0.0
python-magic>=0.4.27; platform_system != "Windows"
python-magic-bin>=0.4.14; platform_system == "Windows"

# Security
cryptography>=40.0.0
certifi>=2022.12.7
pyOpenSSL>=23.0.0

# Utilities
tqdm>=4.65.0
humanize>=4.6.0
pytz>=2023.3
python-dateutil>=2.8.2
typing-extensions>=4.5.0

# Optional dependencies for enhanced features
pillow>=9.5.0  # For thumbnail generation
ffmpeg-python>=0.2.0  # For media processing
beautifulsoup4>=4.12.0  # For web scraping
lxml>=4.9.0  # For XML processing
chardet>=5.1.0  # For character encoding detection

# Development dependencies
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.3.0
isort>=5.12.0
mypy>=1.2.0
flake8>=6.0.0
